+++ tools/DvSchemaSync/Alignment/ColumnAlignmentStatus.cs
@@ -1,8 +1,0 @@
-namespace DvSchemaSync.Alignment;
-
-enum ColumnAlignmentStatus
-{
-    Matched,
-    NativeReplacement,
-    MissingInDataverse
-}