+++ tools/DvSchemaSync/Alignment/KeyAlignmentStatus.cs
@@ -1,8 +1,0 @@
-namespace DvSchemaSync.Alignment;
-
-enum KeyAlignmentStatus
-{
-    Matched,
-    MissingInDataverse,
-    AdditionalInDataverse
-}